Remove getter
It's not generally pythonic to have a "getter" like this - you usually want to just access the object's attributes directly.  It looks like you're signalling to the end user that the width and height aren't part of the object's public API.  I think it's more clean to skip this (and in fact, I'd change the names from _width and _height to width and height).  Others may disagree :)